feat: creamos controllers
